refactor: rename UserRolesTable to UserProjectRolesTable across components
