+++ src/views/administration/accessmanagement/LdapUsers/UserDetails.vue
@@ -50,7 +50,7 @@
     <b-row class="expanded-row p-3" colspan="2">
       <div class="" style="width: 100%">
         <label>{{ this.$t('message.project_roles') }}</label>
-        <user-roles-table
+        <user-project-roles-table
           :parentContext="{ row, index }"
           :projectRoles="projectRoles"
           :availableRoles="availableRoles"
@@ -75,7 +75,7 @@
 import i18n from '../../../../i18n';
 import userManagementMixin from '../../../../mixins/userManagementMixin';
 import ActionableListGroupItem from '../../../components/ActionableListGroupItem.vue';
-import UserRolesTable from '../../../components/UserRolesTable.vue';
+import UserProjectRolesTable from '../../../components/UserProjectRolesTable.vue';
 import SelectPermissionModal from '../SelectPermissionModal.vue';
 import SelectTeamModal from '../SelectTeamModal.vue';
 
@@ -96,7 +96,7 @@ export default {
     ActionableListGroupItem,
     SelectTeamModal,
     SelectPermissionModal,
-    UserRolesTable,
+    UserProjectRolesTable,
   },
   data() {
     return {
